+++ examples/run_classification.py
@@ -95,7 +95,7 @@ def explain_white_box(args):
         # target_layer="last_conv_node_name",  # target_layer - node after which XAI branch will be inserted
         target_layer="/backbone/conv/conv.2/Div",  # OTX mnet_v3
         # target_layer="/backbone/features/final_block/activate/Mul",  # OTX effnet
-        embed_scale=True, # True by default.  If set to True, saliency map scale (0 ~ 255) operation is embedded in the model
+        embed_scaling=True, # True by default.  If set to True, saliency map scale (0 ~ 255) operation is embedded in the model
         explain_method=xai.Method.RECIPROCAM,  # ReciproCAM is the default XAI method for CNNs
     )
 
@@ -309,7 +309,7 @@ def insert_xai_w_params(args):
     insertion_parameters = ClassificationInsertionParameters(
         target_layer="/backbone/conv/conv.2/Div",  # OTX mnet_v3
         # target_layer="/backbone/features/final_block/activate/Mul",  # OTX effnet
-        embed_scale=True,
+        embed_scaling=True,
         explain_method=xai.Method.RECIPROCAM,
     )