+++ openvino_xai/inserter/parameters.py
@@ -21,14 +21,14 @@ class ClassificationInsertionParameters(InsertionParameters):
 
     :parameter target_layer: Target layer (node) name after which the XAI branch will be inserted.
     :type target_layer: str
-    :parameter embed_scale: If set to True, saliency map scale (0 ~ 255) operation is embedded in the model.
-    :type embed_scale: bool
+    :parameter embed_scaling: If set to True, saliency map scale (0 ~ 255) operation is embedded in the model.
+    :type embed_scaling: bool
     :parameter explain_method: Explain method to use for model explanation.
     :type explain_method: openvino_xai.common.parameters.Method
     """
 
     target_layer: str | None = None
-    embed_scale: bool | None = True
+    embed_scaling: bool | None = True
 
     explain_method: Method = Method.RECIPROCAM
 
@@ -44,8 +44,8 @@ class DetectionInsertionParameters(InsertionParameters):
     :type num_anchors: List[int]
     :parameter saliency_map_size: Size of the output saliency map.
     :type saliency_map_size: Tuple[int, int] | List[int]
-    :parameter embed_scale: If set to True, saliency map scale (0 ~ 255) operation is embedded in the model.
-    :type embed_scale: bool
+    :parameter embed_scaling: If set to True, saliency map scale (0 ~ 255) operation is embedded in the model.
+    :type embed_scaling: bool
     :parameter explain_method: Explain method to use for model explanation.
     :type explain_method: Method
     """
@@ -53,7 +53,7 @@ class DetectionInsertionParameters(InsertionParameters):
     target_layer: List[str]
     num_anchors: List[int] | None = None
     saliency_map_size: Tuple[int, int] | List[int] = (23, 23)
-    embed_scale: bool = True
+    embed_scaling: bool = True
 
     explain_method: Method = Method.DETCLASSPROBABILITYMAP