+++ tests/unit/methods/white_box/test_white_box_method.py
@@ -38,7 +38,7 @@ class TestActivationMap:
 
         assert xai_method.model_ori == self.model
         assert isinstance(xai_method.model_ori, ov.Model)
-        assert xai_method.embed_scale
+        assert xai_method.embed_scaling
         assert not xai_method.per_class
         assert xai_method._target_layer == self.target_layer
 
@@ -97,7 +97,7 @@ class TestReciproCAM:
 
         assert reciprocam_xai_method.model_ori == self.model
         assert isinstance(reciprocam_xai_method.model_ori, ov.Model)
-        assert reciprocam_xai_method.embed_scale
+        assert reciprocam_xai_method.embed_scaling
         assert reciprocam_xai_method.per_class
         assert reciprocam_xai_method._target_layer == self.target_layer
 
@@ -156,7 +156,7 @@ class TestViTReciproCAM:
 
         assert reciprocam_xai_method.model_ori == self.model
         assert isinstance(reciprocam_xai_method.model_ori, ov.Model)
-        assert reciprocam_xai_method.embed_scale
+        assert reciprocam_xai_method.embed_scaling
         assert reciprocam_xai_method.per_class
         assert reciprocam_xai_method._target_layer == self.target_layer
 
@@ -233,7 +233,7 @@ class TestDetProbMapXAI:
 
         assert detection_xai_method.model_ori == self.model
         assert isinstance(detection_xai_method.model_ori, ov.Model)
-        assert detection_xai_method.embed_scale
+        assert detection_xai_method.embed_scaling
         assert detection_xai_method.per_class
         assert detection_xai_method._target_layer == self.target_layer
         assert detection_xai_method._num_anchors == self.num_anchors